Update flask_api.py
These changes should update the translation section to be in line with what the professor wanted. If the string is in English, it'll return the string with a dictionary of translations into five different languages. If it is in a language that is already translated in the dictionary, it'll return the string and the English translation. If it is in a non-listed language, it'll return the string and the English translation, but says it is in a non-listed language.